+++ src/components/NavBar.css
@@ -1,0 +1,5 @@
+.gobe-logo {
+    display: inline-block;
+    width: 80px;
+    margin-top: 50px;
+}
