--- src/components/style.css
@@ -1,185 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=IBM+Plex+Sans&display=swap');
-
-#root {
-    overflow-x: hidden;
-}
-
-.main {
-    height: 100%;
-    width: 100%;
-    display: flex;
-}
-
-.gobe-logo {
-    display: inline-block;
-    width: 80px;
-    margin-top: 50px;
-}
-
-.value-proposition {
-    font-family: IBM Plex Sans;
-    align-items: flex-start;
-    margin-top: 10%;
-    font-size: 2vw;
-    font-weight: 900;
-    letter-spacing: 0.020em;
-    box-sizing: border-box;
-}
-
-.welcome {
-    font-family: IBM Plex Sans;
-    font-weight: 900;
-    font-size: 60px;
-    letter-spacing: -0.060em;
-    margin-top: 10%;
-}
-
-.illustration {
-    display: relative;
-    width: 100%;
-    margin-top: 122px;
-}
-
-.welcome-illustration {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    margin-top: 10%;
-}
-
-#signUpForm {
-    margin-top: 1em;
-}
-
-.check-mail {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-}
-
-.confrimation-box {
-    position: flex;
-    font-family: IBM Plex Sans;
-    font-size: 18px;
-    margin-left: 1em;
-    font-weight: 400;
-    /* margin-top: 5%; */
-}
-
-.confirmation-box__div {
-    /* margin-top: 24px; */
-    width: 1.3em;
-    height: 1.3em;
-    /* border-radius: 50%; */
-    background-color: white;
-    border: 1px solid white;
-    vertical-align: middle;
-    border: 1px solid black;
-    -webkit-appearance: none;
-    outline: none;
-    cursor: pointer;
-}
-
-.confirmation-box__div:checked {
-    background-color: #22B573;
-}
-
-.input-box {
-    font-family: Arial, Helvetica, sans-serif;
-    font-size: 18px;
-    border-radius: 30px;
-    margin-top: 1em;
-    border: 2px white;
-    background: white;
-    box-shadow: 0 4px 2px 1px rgb(182, 180, 180);
-    padding: 20px;
-    width: 80%;
-    height: 50px;
-    outline: none !important;
-}
-
-.button {
-    background-color: #22B573;
-    border-radius: 30px;
-    border: 3px #22B573;
-    width: 80%;
-    height: 50px;
-    font-family: IBM Plex Sans;
-    text-align: center;
-    color: white;
-    outline: none;
-    margin-top: 5%;
-}
-
-.error-image {
-    display: flex;
-    width: 70%;
-    margin-top: 10%;
-    margin-bottom: auto;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.centre {
-    position: flex;
-    margin-left: 40%;
-}
-
-.go-back-link {
-    font-family: IBM Plex Sans;
-    font-size: 18px;
-}
-
-.error-message {
-    color: red;
-    font-family: IBM Plex Sans;
-    margin-top: 1em;
-    font-size: 18px;
-}
-
-@media(max-width: 767px) {
-    .illustration {
-        display: none;
-    }
-
-    
-    .main-container__relative {
-        min-height: 85vh;
-        position: absolute;
-        width: 100%;
-    }
-
-    .value-proposition {
-        font-family: IBM Plex Sans;
-        align-items: flex-start;
-        margin-top: 10%;
-        font-size: 6vw;
-        font-weight: 900;
-        letter-spacing: 0.020em;
-        box-sizing: border-box;
-    }
-
-    .main-container {
-        position: relative;
-        background-image: url('../assets/mobile-illustration.svg');
-        background-repeat: no-repeat;
-        min-height: 85vh;
-        background-size: 200%;
-        background-position: center center;
-        background-position-y: bottom;
-    }
-
-    .welcome-illustration {
-        display: none;
-    }
-
-    .welcome-container {
-        position: relative;
-        background-image: url('../assets/xmas_illustration_mobile.svg');
-        background-repeat: no-repeat;
-        background-size: 150%;
-        min-height: 80vh;
-        background-position: center center;
-        background-position-y: bottom;
-    }
-}
