+++ src/pages/NotFoundPage.js
@@ -3,15 +3,16 @@ import { Link } from 'react-router-dom';
 import { HOMEPAGE } from '../routes/urlRoutes';
 import pageNotFound from '../assets/Page-404.svg';
 
-import './style.css';
+import './NotFoundPage.css';
+import { Container } from 'react-bootstrap';
 
 
 const NotFoundPage = () => {
     return (
-        <div>
-            <Link className="go-back-link" to={HOMEPAGE}>Go Back Home</Link>
+        <Container>
+            <Link className="go-back" to={HOMEPAGE}>Go Back Home</Link>
             <img src={pageNotFound} alt="404" className="error-image"/>
-        </div>
+        </Container>
     );
 }
