+++ src/pages/Welcome.css
@@ -1,0 +1,14 @@
+.welcome {
+    font-family: IBM Plex Sans;
+    font-weight: 900;
+    font-size: 60px;
+    letter-spacing: -0.060em;
+    margin-top: 10%;
+}
+
+.welcome-illustration {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    margin-top: 10%;
+}
