+++ src/style.css
@@ -1,0 +1,58 @@
+@import url('https://fonts.googleapis.com/css2?family=IBM+Plex+Sans&display=swap');
+
+#root {
+    overflow-x: hidden;
+}
+
+.main {
+    height: 100%;
+    width: 100%;
+    display: flex;
+}
+
+@media(max-width: 767px) {
+    .illustration {
+        display: none;
+    }
+
+    
+    .main-container__relative {
+        min-height: 85vh;
+        position: absolute;
+        width: 100%;
+    }
+
+    .value-proposition {
+        font-family: IBM Plex Sans;
+        align-items: flex-start;
+        margin-top: 10%;
+        font-size: 6vw;
+        font-weight: 900;
+        letter-spacing: 0.020em;
+        box-sizing: border-box;
+    }
+
+    .main-container {
+        position: relative;
+        background-image: url('./assets/mobile-illustration.svg');
+        background-repeat: no-repeat;
+        min-height: 85vh;
+        background-size: 200%;
+        background-position: center center;
+        background-position-y: bottom;
+    }
+
+    .welcome-illustration {
+        display: none;
+    }
+
+    .welcome-container {
+        position: relative;
+        background-image: url('./assets/xmas_illustration_mobile.svg');
+        background-repeat: no-repeat;
+        background-size: 150%;
+        min-height: 80vh;
+        background-position: center center;
+        background-position-y: bottom;
+    }
+}
